Rewrite so args passed via main method
